refactor: update entity folder structor
