--- src/entities/index.ts
@@ -1,4 +1,0 @@
-export * from '@/entities/user';
-export * from '@/entities/name';
-export * from '@/entities/email';
-export * from '@/entities/password';
