+++ src/entities/user/index.ts
@@ -1,0 +1,1 @@
+export * from '@/entities/user/user-entity';
