+++ src/entities/value-object/email.ts
