+++ src/entities/value-object/index.ts
@@ -1,0 +1,3 @@
+export * from '@/entities/value-object/name';
+export * from '@/entities/value-object/email';
+export * from '@/entities/value-object/password';
