+++ src/entities/value-object/password.ts
