+++ src/use-cases/external-sign-in/external-sign-in-use-case.ts
@@ -1,4 +1,4 @@
-import { User } from '@/entities';
+import { User } from '@/entities/user';
 import { InvalidNameError, InvalidEmailError } from '@/entities/errors';
 import {
   IIdGenerator,
