+++ src/use-cases/sign-in/sign-in-use-case.ts
@@ -1,4 +1,4 @@
-import { User } from '@/entities';
+import { User } from '@/entities/user';
 import { InvalidNameError, InvalidEmailError, InvalidPasswordError } from '@/entities/errors';
 import {
   IHashCompare,
